fix: platform-split CrossWindowDragMonitor + GetAsyncKeyState hover guard
Regression: the 800ms fallback timer fired whenever the cursor hovered
over a native Windows app during a drag, even without releasing the mouse,
creating a spurious new window.

Root cause: the drag/dragleave event-based approaches can't distinguish
"cursor left window (still dragging)" from "mouse released outside window".

Fix:
- Split CrossWindowDragMonitor into platform files (.win32/.darwin/.linux)
  via the existing Vite platformResolve plugin
- win32: dragleave (relatedTarget===null) arms an 800ms timer; when it
  fires, invoke new get_mouse_button_state Tauri command (GetAsyncKeyState
  VK_LBUTTON). If button still pressed → reschedule (user is hovering).
  If released → OLE dropped outside, trigger tearoff.
- darwin/linux: simple dragend-only handler (WebKit/GTK deliver dragend
  reliably, no fallback needed)
- CrossWindowDragMonitor.tsx becomes a platform barrel re-export
