+++ src-tauri/src/lib.rs
@@ -119,6 +119,7 @@ pub fn run() {
             commands::drag::set_drag_cursor,
             commands::drag::restore_drag_cursor,
             commands::drag::release_drag_capture,
+            commands::drag::get_mouse_button_state,
             // File operations
             commands::file_ops::copy_file_to_dir,
         ])
